Payment to User working
Need to change the text in payment sent and payment received in the "Received Money" and "Paid Money" Section in bottom of UI.HTML

+++ register/models.py
@@ -10,7 +10,7 @@ class UserProfile(models.Model):
     currency = models.CharField(max_length=20, choices=[('GBP', 'GBP'), ('Euro', 'EUR'), ('Dollar', 'USD')])
     is_superuser = models.BooleanField(default=False)
 
-    bal = models.IntegerField(default=0)
+    bal = models.DecimalField(max_digits=10, decimal_places=2, default=0)
 
 
     def __str__(self):
